+++ .eslintrc.js
@@ -295,6 +295,11 @@ module.exports = {
                 message:
                   'Use root-based imports (`import "@/something"`) instead of relative imports.',
               },
+              {
+                group: ["../*"],
+                message:
+                  'Use root-based imports (`import "@/something"`) instead of relative imports.',
+              },
             ],
           }),
         ],
