+++ src/components/InvalidatedContextGate.tsx
@@ -28,7 +28,11 @@ const InvalidatedContextGate: React.FunctionComponent<{
   // Only auto-reload if the document is in the background
   const isDocumentVisible = useDocumentVisibility();
   if (wasContextInvalidated && autoReload && !isDocumentVisible) {
-    location.reload();
+    setTimeout(() => {
+      // If you reload too soon, Chrome might not be ready to serve the page yet
+      // TODO: Poll the page until it's ready instead of a timeout. Then auto-reload by default
+      location.reload();
+    }, 500);
   }
 
   return wasContextInvalidated ? (
