+++ src/contentScript/messenger/registration.ts
@@ -46,10 +46,8 @@ import {
   runRendererBlock,
 } from "@/contentScript/pageEditor"; // 300 strictNullCheck errors
 import { runBrick } from "@/contentScript/executor"; // 290 strictNullCheck errors
-import {
-  cancelSelect,
-  selectElement,
-} from "@/contentScript/pageEditor/elementPicker"; // 290 strictNullCheck errors
+import { cancelSelect } from "@/contentScript/pageEditor/elementPicker"; // 290 strictNullCheck errors
+import { selectElement } from "@/contentScript/pageEditor/selectElement";
 import {
   runHeadlessPipeline,
   runMapArgs,
