+++ src/contentScript/pageEditor/elementPicker.ts
@@ -18,27 +18,14 @@
 
 import Overlay from "@/vendors/Overlay";
 import ReactDOM from "react-dom";
-import {
-  expandedCssSelector,
-  findContainer,
-  inferMultiElementSelector,
-} from "@/utils/inference/selectorInference";
+import { expandedCssSelector } from "@/utils/inference/selectorInference";
 import { compact, difference, uniq } from "lodash";
-import * as pageScript from "@/pageScript/messenger/api";
-import { type SelectMode } from "@/contentScript/pageEditor/types";
 import {
   type SelectionHandlerType,
   showSelectionToolPopover,
 } from "@/components/selectionToolPopover/SelectionToolPopover";
-import {
-  BusinessError,
-  CancelError,
-  NoElementsFoundError,
-} from "@/errors/businessErrors";
+import { BusinessError, CancelError } from "@/errors/businessErrors";
 import { FLOATING_ACTION_BUTTON_CONTAINER_ID } from "@/components/floatingActions/floatingActionsConstants";
-import { $safeFind, findSingleElement } from "@/utils/domUtils";
-import inferSingleElementSelector from "@/utils/inference/inferSingleElementSelector";
-import { type ElementInfo } from "@/utils/inference/selectorTypes";
 import { onContextInvalidated } from "webext-events";
 
 /**
@@ -421,87 +408,5 @@ export async function userSelectElement({
 export async function cancelSelect() {
   if (_cancelSelect) {
     _cancelSelect();
-  }
-}
-
-export async function selectElement({
-  mode = "element",
-  root,
-  isMulti: initialIsMulti = false,
-  excludeRandomClasses,
-}: {
-  mode: SelectMode;
-  isMulti?: boolean;
-  root?: string;
-  excludeRandomClasses?: boolean;
-}): Promise<ElementInfo> {
-  const rootElements = $safeFind(root).get();
-
-  if (root && rootElements.length === 0) {
-    throw new NoElementsFoundError(root);
-  }
-
-  const { elements, isMulti, shouldSelectSimilar } = await userSelectElement({
-    roots: rootElements,
-    isMulti: initialIsMulti,
-  });
-
-  console.debug("Selected elements", { elements, isMulti });
-
-  switch (mode) {
-    case "container": {
-      if (root) {
-        throw new Error(`root selector not implemented for mode: ${mode}`);
-      }
-
-      const { selectors } = findContainer(elements);
-
-      findSingleElement(selectors[0]);
-
-      return pageScript.getElementInfo({
-        selector: selectors[0],
-      });
-    }
-
-    case "element": {
-      let activeRoot: HTMLElement | null = null;
-
-      if (isMulti) {
-        // If there are rootElements, the elements must all be contained within the same root
-        activeRoot = rootElements?.find((rootElement) =>
-          elements.every((element) => rootElement.contains(element)),
-        );
-
-        return inferMultiElementSelector({
-          elements,
-          root: activeRoot,
-          excludeRandomClasses,
-          shouldSelectSimilar,
-        });
-      }
-
-      if (elements.length !== 1) {
-        console.warn(
-          "Expected exactly one element for single element selector generation",
-        );
-      }
-
-      const element = elements[0];
-      // At least one much match, otherwise userSelectElement would have thrown
-      activeRoot = rootElements?.find((rootElement) =>
-        rootElement.contains(element),
-      );
-
-      return inferSingleElementSelector({
-        root: activeRoot,
-        element,
-        excludeRandomClasses,
-      });
-    }
-
-    default: {
-      const exhaustiveCheck: never = mode;
-      throw new Error(`Unexpected mode: ${exhaustiveCheck}`);
-    }
   }
 }
